[dot net v2] Created 2 projects for dot net 40 and dot net 45
