--- src/Payment.aspx.cs
@@ -1,33 +1,0 @@
-﻿using Razorpay.Api;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.UI;
-using System.Web.UI.WebControls;
-
-namespace RazorpaySampleApp
-{
-    public partial class Payment : System.Web.UI.Page
-    {
-        public string orderId;
-        protected void Page_Load(object sender, EventArgs e)
-        {
-            Dictionary<string, object> input = new Dictionary<string, object>();
-            input.Add("amount", 100); // this amount should be same as transaction amount
-            input.Add("currency", "INR");
-            input.Add("receipt", "12121");
-            input.Add("payment_capture", 1);
-
-            string key = "Your Key";
-            string secret = "Your Secret";
-
-            RazorpayClient client = new RazorpayClient(key, secret);
-
-            Razorpay.Api.Order order = client.Order.Create(input);
-            orderId = order["id"].ToString();
-
-
-        }
-    }
-}
